update python module

+++ src/mathUtils.h
@@ -3,7 +3,7 @@
 
 inline float sigmoid(float x)
 {
-	return 1 / (1 + exp(-x));
+	return 1 / (1 + expf(-x));
 }
 
 struct F_logsigmoid
